Fetch actual NVI periods instead of using hardcoded list
