+++ src/types/nvi.types.ts
@@ -109,6 +109,7 @@ export interface Note {
 
 export interface NviPeriod {
   type: 'NviPeriod';
+  id?: string;
   publishingYear: string;
   reportingDate: string;
   startDate: string;
